+++ lib/functions/is-exact/index.ts
@@ -1,0 +1,6 @@
+import { Exact } from "../../exact";
+
+export const isExact =
+  <ExpectedShape>() =>
+  <ActualShape>(x: Exact<ActualShape, ExpectedShape>) =>
+    x as ExpectedShape;