+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # ts-essentials
 
+## 9.4.2
+
+### Patch Changes
+
+- f88f757: Add TypeScript 5.3 and 5.4 support (fix a bug with `WeakKey` for `WeakSet` and `WeakMap`)
+
+## 9.4.1
+
+### Patch Changes
+
+- bc3c474: Fix regression in TypeScript 5.1 with CamelCase
+
+## 9.4.0
+
+### Minor Changes
+
+- ddb1bea: Make `typescript` optional as a peer dependency
+
+## 9.3.2
+
+### Patch Changes
+
+- 8761667: `DeepPartial`, `DeepRequired`, `DeepWritable` and `DeepReadonly` can update structure of `Error` or interface
+  which extends `Error`
+- afccd35: Fix handling of readonly arrays in `isTuple`, `DeepNullable`, `DeepPartial`, and `DeepUndefinable`
+
+## 9.3.1
+
+### Patch Changes
+
+- 0c215b9: Changed the `Tuple`'s type implementation to only use a tuple type (instead of a union of a tuple type and an
+  array type).
+- 544742e: Fix unique symbol `__OPAQUE_TYPE__` in `Opaque` that cannot be used outside of ts-essentials
+- 4c45165: Improve the object constraint for `NonEmptyObject<T>` to not allow primitives
+- 66a6169: Improve the `ValueOf` utility type to cover the edge cases
+
 ## 9.3.0
 
 ### Minor Changes